entity for company metrics, team & investment added

+++ src/main/java/com/unicorn/studio/dao/InvestorRepository.java
@@ -3,6 +3,6 @@ package com.unicorn.studio.dao;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.unicorn.studio.entity.Investor;
 
-public interface InvestorRepository extends JpaRepository<Investor, Integer> {
+public interface InvestorRepository extends JpaRepository<Investor, Long> {
 
 }
